JSObject 相关类接口化改造
