+++ wrapper-java/src/main/java/com/whl/quickjs/wrapper/QuickJSFunction.java
@@ -1,0 +1,20 @@
+package com.whl.quickjs.wrapper;
+
+/**
+ * Created by Harlon Wang on 2024/2/12.
+ */
+public class QuickJSFunction extends QuickJSObject implements JSFunction {
+
+    private final long thisPointer;
+
+    public QuickJSFunction(QuickJSContext context, long pointer, long thisPointer) {
+        super(context, pointer);
+        this.thisPointer = thisPointer;
+    }
+
+    @Override
+    public Object call(Object... args) {
+        return getContext().call(this, thisPointer, args);
+    }
+
+}
